requirements.txt
Django==5.0
djangorestframework==3.16.1
pillow>=7.1.0,<12.0
python-decouple==3.8
psycopg2-binary==2.9.11
requests==2.32.5
whitenoise==6.11.0
django-cors-headers==4.9.0
pytest-django==4.11.1
gunicorn==21.2.0
openai==1.57.4
dj-database-url==2.1.0
